Remove specific deployment extra classpath and isolation group that were only available in Java 8
